feat(util): add themeResolver (#69)
This adds a themeResolver util that converts a ThemeConfig to a
TailwindConfig format, and combines it with the developer-specified
tailwind config that is also passed in. Configurations made by the
developer take priority over the user/marketer's theme config. Tested by
building locally and using the local build in the starter. I put part of
the tailwind config into theme.config.ts and passed both the theme
config and the rest of the tailwind config into themeResolver, which
provided the same output as the old tailwind config.