feat(util): add buildCssOverridesStyle (#72)
This adds a `buildCssOverridesStyle` util function that takes in an
object containing the variable name and value overrides that need to be
added to the page in a `style` tag. The `selector` field will be
`.components` for our use case but I left it as a parameter in case we
want to reuse it later. It also takes in an optional `base` field that
gets prepended to the string that gets returned, in case the user needs
to add anything else to the page's `<head>` tag besides these styling
overrides.

Verified the function works via the test cases and running it manually
and confirming the returned string is what I expected.