feat(components): add react-color color picker (#68)
Verified that the color picker works and is interactable in the theme
sidebar.

Tested in dev mode

ColorSelector can be added to props in components and will then appear
in the editor. Colors can be edited and the state persists through
reloading. Color is reset when clearing local data.

Example addition to Hero.tsx in
`github.com/yextSolutions/pages-visual-editor-starter`
```
  myColor: ColorSelector({
    label: "My Color",
    type: "custom",
  }),
```

Which appears in the editor as:


![image](https://github.com/user-attachments/assets/b4fa997b-ac12-4abf-a0ae-9c54faaa127a)

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>