fix: update puck history with new layout's config in theme mode (#93)
While in theme mode, this sets the puck history to the updated
`puckInitialHistory`, which is derived from the visualConfigurationData.
This ensures that the editor updates accordingly when a different layout
is selected.

Tested manually by running a local build of visual-editor with the
starter. Changing the layout in the header made the editor update with
the new layout after about half a second.