fix: error in constant text fields (#199)
Fixes a bug I introduced in #193 
Was trying to use the custom ConstantField styling for all constant
value inputs, but missed that the ConstantField function depended on the
the field being an object type. This PR reverts Phone and Text to being
text types but adds a case to render a styled FieldLabel for them.