fix: resolveData stuck when updating footer props (#200)
getFontWeightOverrideOptions is set up to wait for the DOM to mutate
before getting the theme data because puck runs resolveFields before the
document is fully loaded.

However, if getFontWeightOverrideOptions is used in a manually-defined
resolveFields function (like I added in footer), resolveFields hangs
because the DOM doesn't mutate further. This results in the sidebar
being stuck in a loading state.

This PR updates getFontWeightOverrideOptions to try to get the theme
data when it is called and then start the MutationObserver if it cannot
be found.