refactor: remove uuid from Header and Footer (#201)
Originally added for the React key to work properly but I guess the
original issue was something with combining the existing footer with
arrays from the Puck demo because it doesn't seem necessary.

This PR simplifies Header/Footer by just using the array index as the
key, and it seems to work totally fine with
add/remove/duplicate/reorder.