fix: format phone numbers correctly (#227)
This change makes the Phone component correctly format phone numbers
based on the country code. This uses the `libphonenumber-js` library to
automatically parse the phone number and then format it in either the
domestic or international format. Tested locally and ensured that phone
numbers with various country codes were formatted as expected.