feat: add page level fields for Title and Description (#225)
Adds page level fields for title and description in the puck editor.

They can use entity values or constant values.

They are set to use default values pointing to the name and description
fields.

Also adds a new exported function getPageMetadata(). This will return
the title, description, and any other key/value pairs from root.

Note: the history does not update when the root is updated. However if
any other fields are updated, then changes to root will be included in
the history. This means that changing the root alone will not save to
saveState or enable the Publish and Clear Local Changes buttons. However
if anything else is changed, then root changes will be
saved/published/cleared as-well.

---------

Co-authored-by: Matt Kilpatrick <mkilpatrick@yext.com>