fix: fixes root props not updating or saving properly
If root.props is missing from the history state, then root prop
changes do not trigger updates to histories or index from usePuck.

By ensuring that there are props on the history state, we ensure
that the root props update properly.

Verified that root prop changes update and save properly in
local dev mode and in production mode.