fix: useEntityFields must be used within VisualEditorProvider (#239)
usePlatformBridgeEntityFields was undefined until the platform sent down
the stream information, which would occasionally happen after all the
other data was sent down. The other data was checked before loading the
editor but this wasn't because it's set as context outside of the normal
loading flow. `useEntityFields must be used within VisualEditorProvider`
was thrown when the entityFields are undefined and useEntityFields is
called, which is now immediately after loading because we render a
YextEntityField selector for the root fields.

This fix distinguishes between not-in-provider and no-loaded-yet and
checks that entity fields have loaded before loading the editor