refactor: update components to use Link (#244)
Updates GetDirections to use CTA instead of a Button + Link combo.

Updates various other components to add a Link Type prop or set one by
default.

Fixes a VE crash when entering a blank constant value for Promo or Card
CTA link.

Verified that the appearance and functionality of the components are not
any different than before.