feat: add includeHyperlink option to Phone component (#246)
Adds an option to include a hyperlink for the phone component.

The option is similar to what is available in the Emails component.

When a hyperlink is included, an underline is added to the phone that
disappears on hover.

<img width="1397" alt="image"
src="https://github.com/user-attachments/assets/0ce6ac82-ecdb-480d-896e-d24b098e97f0"
/>