refactor: new theme config (#282)
Aligns the colors, headings, body and page sections of the theme config
with the document. Also updates of the heading variables.

Usages of colors and body styles will need to be fixed in a future PR
following the addition of auto-gen colors and body styles (I will work
on that next). The section component (and Flex and Grid) will need some
updates as well, since we no longer set gap and have different padding
types.