fix: handle non-string values correctly in BasicSelector (#289)
cmdk's CommandItem component expects all values to be strings, and
produces inconsistent results when non-string values are provided. We
want our BasicSelector to work for non-string values like numbers (such
as in the HeadingText component's level field), so this change
stringifies the values that are passed in, and updates BasicComponent to
select the option with the value that, when stringified, matches the
value that was clicked. This also updates the cmdk package to ensure we
get other bug fixes that have been released (although the newest version
still does not handle non-string values).

Tested manually and confirmed that BasicSelector works as expected for
both string and non-string values.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>