refactor: update text and background color for components (#286)
Uses the new background styles for all background and text color

In general, removes text color as a class that is set in individual
components, so that the text color set by the Grid/Flex/Section is
inherited.

CTAs still need to be updated to the new styles, so they are a little
off.



https://github.com/user-attachments/assets/f3b2b655-1927-47b1-95c1-bb828b9f0661