feat: integrate mapbox map component within locator component
I have added the MapboxMap from search-ui-react to the Locator component.
Currently, the only property of the map that is configurable is the "style"
of the map (e.g. street, dark mode, day navigation, etc).
The map will attempt to find the user's location, and use that as the
first search. The user can either drag/zoom around the map and click
"Search this area" to run a search, or they can use the filter search bar.

The display will be the following:
- for small screen sizes, the map component will be hidden
- for medium and larger screen sizes, the filter search + results will
occupy the left 1/3 of the screen, while the map will occupy the remaining
right 2/3.

To see it in action, see this video:
https://drive.google.com/file/d/1hvAdD6mtwnOLRXfU_0cN0jokzBqnBfOx/view?usp=drive_link

J=WAT-4700