feat: update grid component (#294)
This removes the Maximum Content Width option from the Grid component,
and simplifies the behavior of columns. The column field is now a basic
number field, and columns wrap to rows on small screen sizes if the Grid
is at the top level of the layout, but are fixed to a horizontal
alignment at all times if the Grid is at a lower level. This also
renames the padding fields and shows the raw px values for spacing
rather than the more complicated "[number] ([pixel_value]px)" format.

This screenshot shows two Grids containing CTA buttons on a small mobile
screen. The top grid is at the top level of the layout, so it wraps into
one column with multiple rows. The bottom grid is within another grid,
so the buttons are always horizontally aligned into columns.

<img width="994" alt="Screenshot 2025-03-26 at 11 51 35 AM"
src="https://github.com/user-attachments/assets/dd5f4570-2bb4-4c9e-9f72-58e1094e4d6a"
/>