fix: image import build error (#308)
Was getting `[vite:css] [postcss] Cannot access 'Ur' before
initialization` -- `Ur` is the minified name of `ImageWrapperFields`

I suggested to move the ImageWrapper imports to be from the package in
#300 but that was a bad suggestion. Basically, you can't do
`../../index.js` imports if the object you are importing is exported at
the same level as the file that is importing it. The declaration is
hoisted but it's not been initialized yet.

I think we can just do the direct import for now since we're probably
adding a YextField helper soon