refactor: button/link/cta color handling (#313)
I've switched from finding the actual rendered background color via the
DOM to setting background color in context. The original method occurred
after render and so would cause flashing on the page. The new method is
slightly less ideal in that it requires intentionally setting the
context in a parent rather that it working on its own. However, almost
everything that sets a background color should use `<Section>` which
will handle it for you, and I've exposed BackgroundProvider if you want
to do something custom.