get width and height dynamically
There is a bit of a chicken and egg problem

Before API results are returned, the parent width is tiny because
the component takes up no space.

If we then set the width value to the width of the parent, we get a very
thin map returned.

Setting a min size of 300 makes the map fill out components properly and
looks pretty good.

We could expose a min width prop, or get the grandfather component and
try to figure out what the width should be to fit in the collection.

When placed in a section, it uses the original size params of 1024x300.

In a flex, it will end up with the original size of 1024x300.