chore: add wcag testing (#367)
Adds wcag tests for each component separately and with all the
components on the page, using `vitest`, `jest-axe`, `playwright`, and
`axe-core`. In cases where the default props do not render a component,
test data has been added too. We may still want some way of testing the
default layout but that's stored in the starter. When run locally,
screenshots are taken of each test for debugging