fix: hero link misalignment (#401)
In the Hero component, we have two CTAs

If one is set to Link and the other a button style, then the Link will
appear offset to the bottom.

This adds "py-3" to both of the CTAs to ensure that they both have it,
even when in "Link" style.

This makes things stay aligned without affecting buttons in other
places.

Fixed: 

![image](https://github.com/user-attachments/assets/0599fccc-d86c-4100-b48f-61718d373c3f)

Broken:

![image](https://github.com/user-attachments/assets/89044f5c-c10f-47ac-88e6-e09a79d16242)