feat: merge repo structure updates into locator dev branch (#410)
This merges the new changes in the last few weeks into the Watson dev
locator branch.

Additionally, some of the recent changes have caused the Locator to have
issues scaling in interactive mode, due to the height now being `100vh`
(set in `starter/src/ve.config.tsx`). I have adjusted the `Locator`
component to now have a height of `fit-content` to address this. I have
also adjusted the width for the results / map in the medium viewport, to
now be 2/5 & 3/5, rather than 1/3 & 2/3.

---------

Co-authored-by: Alexis Sanehisa <106991746+asanehisa@users.noreply.github.com>
Co-authored-by: Ben Life <77246839+benlife5@users.noreply.github.com>
Co-authored-by: Duval Kilpatrick <mkilpatrick@yext.com>
Co-authored-by: Colton Demetriou <88292188+colton-demetriou@users.noreply.github.com>
Co-authored-by: briantstephan <70166284+briantstephan@users.noreply.github.com>
Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Ben Life <blife@yext.com>
Co-authored-by: Rakesh Keerthi <108081082+rkeerthient@users.noreply.github.com>
Co-authored-by: Jacob <115099870+jwartofsky-yext@users.noreply.github.com>
Co-authored-by: Fondryext <160865254+Fondryext@users.noreply.github.com>
Co-authored-by: Jacob Fondriest <jfondriest@yext.com>