feat: support fullstory in editor iframe (#423)
Adds fullstory capture to the edit page

Testing: Created a branch with a self-packed version of VE. Swapped to
that branch on a prod account that is not excluded from reporting.
Logged out of admin2 and then into the platform as a user created from
the email signup flow. After waiting at least 30 minutes for my
fullstory session to populate, checked with Duval (who has access to
fullstory) and confirmed the iframed content was visible)