feat: add YextStructFieldSelector and update hero (#432)
https://jam.dev/c/0646a03b-d27f-4c8a-89bd-7ea44559d8ce

Made a new YextStructFieldSelector for now rather than attempt to make
it with YextEntityFieldSelector. This new functionality with constValue
overrides was requested at a recent meeting but would like to wait for
us to see if we like it before cleaning up YextEntityFieldSelector and
merging them (if we merge).

For HeroSection we wanted 3 main fields, data, styles, and
liveVisibility. We also purposefully removed the showHide fields for the
ctas and hours.