feat: create locator template (#433)
Creates the template for locator page sets. This is very similar to the
directory template, however it differs in that it uses the search
analytics config rather than the pages component, and imports the
`locatorConfig` from `ve.config` instead of the `directoryConfig`.

Additionally, we still need to make the necessary update to the starter
repo's `ve.config` file, as well as updates to this repo to pull the
experience key off of the entity document.

J=[WAT-4782](https://yexttest.atlassian.net/browse/WAT-4782)
TEST=none