chore: improve automated tests (#459)
Each of the PageSections now has 4 tests (run for mobile and desktop):
- default props with empty document
- default props with data in the document
- version 0 props that use as many entity values as possible
- version 0 props that use as many constant values as possible
The rendered component is then tested for WCAG and that each piece of
expected data appears on the page.

Migrations are run before each test so the a component's test should
fail the v0 tests if it has prop changes that are not properly
migration. (Conversely, updating `expect` for a version 0 test indicates
we are making breaking changes).