feat: add migration check (#462)
Adds a github action that checks if component files have been updated.
If they have and migrations have not, a comment is added to the PR with
a warning. Doesn't actually check for anything other than files being
updated so it's far from comprehensive but may be a helpful nudge?

See example here: #461 (the first commit had component changes but no
migration, the second commit component+migration changes, and the third
no component changes)