feat: update analytics (#496)
Update analytics event names for each component such that each event is
uniquely named.

Scope is not available as a field for users to set, but it would be easy
to add
```
analytics: YextField("Analytics", {
    type: "object",
    objectFields: {
      scope: YextField("Scope", {
        type: "text",
      }),
    },
  }),
```

No migrations are required as having a null analytics.scope is handled.
If scope is set to "" in future (if it is settable by users) then no
scope will be used.