feat: use YEXT_EDIT_LAYOUT_MODE_MAPBOX_API_KEY when in VLE if present (#495)
We will use an environment-wide Mapbox key (in the
`YEXT_EDIT_LAYOUT_MODE_MAPBOX_API_KEY` env var) when in the VLE iframe
in order to get around the URL restrictions by the site-specific API
key. If this key is not present for some reason, we will fall back to
the site-specific API key.

J=[WAT-4873](https://yexttest.atlassian.net/browse/WAT-4873)
TEST=manual