feat: add translations for HoursStatus (#494)
Adds overrides for day of the week and the current status.
    
Day of the week can be easily overridden to pass the real locale
through. If alternate locales are passed, the translated weekday
is used.
    
Adds translation files for current status and future status.

Applies the translation to the HoursStatus component and HeroSection
component.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>