feat: add platform translations part 1 (#509)
Adds translations for the editor text that will use the Storm user
locale. Didn't update the insert script because a lot of these were
one-off situations or required a restructure.

The field labels are still set in the components and require a more
complicated translation setup. Looking into that now.


https://github.com/user-attachments/assets/132c528c-53f4-485f-8a87-c0af8e1744a9