feat: add platform translations part 2 (#521)
1. Translates puck labels
- Since the Puck labels are stored as JSON rather than TSX, we can't use
i18next's `t` in the configs. Instead, we have to replace the labels
with the translated version in the various render functions
- I've added a function `msg` that serves as a translation wrapper to
indicate to the i18n scanner that the string should be translated. It
sets the puck label to a JSON-ified set of translation options that are
later run through the `pt` function
2.  Converts entity field labels to use platform translations

There's still a few rough spots that I'm working on or at least trying
to triage what is necessary/easy vs. not worth it for now
- a few instances of string concatenation that should use translation
interpolation
- a few weird/bad responses from Google Translate that context would
help
- field display names from KG



https://github.com/user-attachments/assets/02f1aafa-26ea-40a4-a19f-5ba6ecc00f09