feat: add Language Dropdown to Header (#514)
There is no language dropdown if the pageset has less than two locales.
Once the language dropdown is clicked, the pagesets' locales are
compared to the entity's locales. If the entity has only one locale, the
dropdown doesn't 'disappear' but only shows that one locale. If the
entity has multiple locales also in the scope, those locales are shown
and clicking on changes the href to the new slug.

This only works for in-platform developers.


https://github.com/user-attachments/assets/02e0875b-ce01-41be-a83b-ed12c59bf30e

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>