chore: remove collections (#523)
Puck 0.19 introduces slots which will allow for a much better collection
experience than the existing skateboard. This PR removes the Collection
component, the Card components, and the associated utilities.

If we re-add cards, they will probably be based on the changes we've
made in PageSections and we could always go back in git history if
needed.