feat: edit constant value rtf fields in drawer (#538)
Pops a drawer in Storm that will be used to edit RTF. For now, these
fields just use a simple string.

Although RTF could have been done without storm, the mocks show the
drawer taking the full screen height which can't be done in the iframe,
we get access to Mana, and we're going to need to do this for image
assets anyway.

To Do (probably in separate PRs): 
- Build RTF editor in storm
- Handle storing/passing/rendering RTF
- Handle or potentially remove non translatable RTF?




https://github.com/user-attachments/assets/2bd5cfa1-aca9-49a6-af44-5c96b8676f1d