feat: update i18n for context (#559)
Appears to be working for contexts such as "direction", "text size", and
"phone number" based on my limited understanding of spanish / japanese.

This basically passes it into google api like "Right [[direction']]" and
then removes the context afterwards so we get "Derecha" for spanish's
right (direction) instead of "Bien" which is more like right (correct).