refactor: adjust analytics (#562)
- Fix: CoreInfoSection was missing a scope because it didn't have a
scope in the defaultLayoutData or a fallback value in the render method
- Fix: ProductSection event numbers were undefined because `key` is a
reserved prop that doesn't get passed to the component. Updated in other
places as well
- Feature: Append a 3-digit has to the end of each component scope to
differentiate different components on the page. I hashed the Puck id
because that is the one thing that remains constant for as long as the
component is part of the layout (ie if you move it on the page it will
remain the same, but if you remove and re-add it will be different).

<img width="1840" alt="Screenshot 2025-06-26 at 1 36 48 PM"
src="https://github.com/user-attachments/assets/0151980d-b88b-42d1-a884-cd0a651c4d9a"
/>
<img width="1840" alt="Screenshot 2025-06-26 at 1 36 44 PM"
src="https://github.com/user-attachments/assets/6989d787-47de-495e-b441-4872de878b2d"
/>

Note for future analytics work: if you set `enableDebugging={true}` for
the `AnalyticsProvider` in `main.tsx`, you can use proxied dev mode to
test (thanks Alexis for the tip)