feat: add Open Now button to locator (#569)
This PR adds the ability to add an 'Open Now' button to the
Locator component. If enabled, and if clicked, a filter is
applied to only show the currently open results.

To accomplish this, a new component atom, toggle, has also been added.
For now it has
no props, and always uses the secondary color when toggled.
J=WAT-4908
TEST=manual

Ran in interactive dev mode and saw all expected behavior.
Reference the Jam video
https://jam.dev/c/bb97c575-2e20-4d9e-83c0-0e9594e584c0

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>