chore: clean up header/footer image styling (#598)
![Screenshot 2025-07-09 at 3 14
17 PM](https://github.com/user-attachments/assets/2433c869-c029-4883-aa1f-c23cb188648a)
I can revert the default change, but I thought this looks better than it
going all the way to the right side of the page. (Centering the items
also fixes this, but then if you increase image size it grows towards
the left side of the page instead of to the right).

---------

Co-authored-by: coltondemetriou <coltondemetriou@gmail.com>