feat: revamp grid to new core information grid section
This PR takes the currently unused Grid and revamps it
for usage as the new Core Information grid.
Any of the other subcomponents can be pulled into the grid.
Columns can be set to 2 or 3, but on medium and small screens
the width is only 2 and 1 columns respectively.
J=WAT-4960
TEST=manual

Used the starter repo to test the components and saw them
working as expected