feat: support for component filtering (#673)
- The gated components will only be visible and draggable if a bussiness
has the appropriate product feature enabled.
- If a gated component is present on a page, but the feature is disabled
later, it will not show on live page.
- In the editor, the component will not appear on the left hand side
components list. If it was previously present on the page, a section
with the text "No configuration for XyzSection" will be displayed on the
canvas instead of the gated component. It is removable.

J=WAT-4956
TEST=manual

reference this packed version in a starter branch and tested in
platform. verified that the editor and live page look as expected when
with the product feature enabled/disabled.

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>