fix: revert hoursStyling changes (#692)
Using max width with align right would look strange in certain
situations

I've also noticed that if the services column is missing, then the other
columns properly take up the space.

In this case the CoreInfo column is much wider and with full width it
looks quite odd.

This reverts the HoursStyling changes while keeping the header changes
from:
https://github.com/yext/visual-editor/commit/0396e0fa5078bbfe411e0b153a064e518f173112

Additionally this updates CoreInfoSection to hide the CoreInfo column if
it has no data

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>