refactor: hybrid developer work flow (#699)
This replaces the less intuitive / messier optional props for components
that need values from the hybrid-developer and now lets the
hybrid-developer set them at the top level in their template. Confirmed
with active hybrid developers they were not using the old flow so
deletion is okay.

Tested on my hybrid site and confirmed it works both in the Editor and
on the Live Page.

Edited the hybrid-development.md to explain new workflow.

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>