feat: add CTA Group component (#697)
The CTA Group component will be part of the components hidden by the
advanced feature flag + can be used in Grid.

For the cta atom, I just split it up into two parts, one CTA element + a
useResolvedCtaProps function so the code is a bit more readable (imo).
Tested and confirmed no changes to functionality.

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>