fix: prevent infinite re-renders with PhoneInput (#736)
The new `PhoneInput` component exhibits weird behavior that our old one
did not. When selecting a new country code with the dropdown, it would
repeatedly swap back and forth between the new country code and the old
phone number. This seems to be caused by how it interacts with the
`value` that is passed into `PHONE_CONSTANT_CONFIG`'s renderer. Managing
the value in an internal state fixes this issue, and adding a hidden
dummy button before the `PhoneInput` prevents the dropdown from
re-opening after selecting a new country code.


https://github.com/user-attachments/assets/d25a6dab-04e7-4e69-b7e5-b968ff9a3c8e

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>