fix: category filtering (#753)
In the latest artifact, the Custom Code Section appears in the sidebar
even with the feature flag off however you cannot drag it onto the page.
Tested this change in platform dev mode and confirmed that the feature
flags worked as expected.

I thought this used to work but here's an artifact 0.0.17 site, which is
when we introduced the Custom Code Component, and it shows up even with
the feature flag off:
https://dev.yext.com/s/1000152098/yextsites/65518/editor#pageSetId=locations&locale=en&themeId=artifact-0-0-20&entityId=8725521

We might want to release a new artifact since this is affecting GA.

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>