fix: dark brand colors in RTF (#760)
I missed this background in #631. This makes RTF work with the dark
brand colors (it already worked with the dark recommended backgrounds).
Also, BodyText wasn't updated then because it wasn't in use

BodyText is still not ideal (we should be able to pass props like
variant into resolveComponentData) but that's a bigger fix:
https://yexttest.atlassian.net/browse/SUMO-7762

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>