fix: don't treat advanced settings as a component (#772)
<img width="598" height="369" alt="Screenshot 2025-09-22 at 11 02 34 AM"
src="https://github.com/user-attachments/assets/0c893c05-849b-430f-a82c-d49426009e08"
/>

treating advanced settings as a component was breaking drag and drop
functionality for other components, it also appeared in the page outline
which was not ideal. We will need to use a back button here given we
cannot have breadcrumbs without treating it as a component.

---------

Co-authored-by: coltondemetriou <coltondemetriou@gmail.com>