feat: add heading text slot (#790)
Updates EventSection to use a slot component for the section heading
text. Refactors HeadingText to include an align prop and organize props
into data/styles

Will update the rest of EventSection (+tests) in separate PR but heading
will be a shared component

Used a SlotsCategory because it seems like `resolvePermissions` doesn't
support checking if the component is in a slot or not. Will start a
thread with the Puck team because that would be easier.


https://github.com/user-attachments/assets/b9758171-3f8e-4ba8-8f31-1ee1d18ceb26

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>