fix: fonts not loading in editor (#793)
<img width="1221" height="999" alt="Screenshot 2025-10-01 at 1 59 29 PM"
src="https://github.com/user-attachments/assets/80cdc2f4-da2f-413b-ad1d-6ac4ba5263d1"
/>
Confirmed that fonts now load correctly in both the Theme and Layout
Editor and match the live page.

The problem was that the live page used applyTheme but the editor was
only using updateThemeInEditor, so when the font was changed in the
editor, the new styles were applied but the actual font files weren't
loaded, leading to the editor falling back to defaults. The fix was
essentially to just use the applyTheme logic in the editor as well.

Test Site:
https://www.yext.com/s/4259018/yextsites/160744/theme#themeId=font-test-2&pageSetId=font-test-2&locale=en&entityId=1095527849

---------

Co-authored-by: coltondemetriou <coltondemetriou@gmail.com>
Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>