feat: provide default schema to parent window (#791)
When sending the user's schema to the parent to allow for editing, we
now also send the default schema. A new option in the UI will allow the
user to restore the default schema and clear any edits they've made.

Ran everything locally and logged the payload that was sent to the
front-end, confirming that it included the new defaultValue field.

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>