refactor: cta prop fields  (#819)
We added the `ctaType` field with three options: `textAndLink`,
`getDirections`, and `presetImage`. This was originally added to the
`constantValue` object but later also tied into entity field selection.
In the current artifact, these options are only available for Hero/Promo
and the content blocks, but going forward every slot component will use
CTAWrapper.

There were a few bugs
- the entity field selector doesn't update when swapped to
`getDirections`
-  the image for `presetImage` is not selectable in entity mode
- CTA variant (primary/secondary/link) was always prop even though it
does not apply to `presetImage` mode

A lot of the complexity comes from that `presetImage` is really a
styling variant of `textAndLink`, whereas `getDirections` is a different
data type. This change:
- moves the option for `textAndLink` vs. `presetImage` to the `styles`
section of CTAWrapper
- `variant` and the image for `presetImage` are styles fields that are
conditionally hid/shown
- visually moves the `link` vs `getDirections` option in the `data`
section to the top of the header
- moves the `link` vs `getDirections` option in data payload out of
`constantValue` since it applies to both entity values and constant
values
- ensures the entity field selector always updates when `link` vs
`getDirections` is toggled
- hides CTAs that don't have a label, coordinate, or preset image

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>