Add facets to locator component filter modal
Updates the locator component's filter modal to include the facets selected by the user in the settings panel.

Note: A know gap between this commit and the mock is that the checkboxes are gray rather than the primary themem color (in the case of the demo, dark red) is a know weakness of this stage. That will require additional work in search-ui-react to allow for custom icons for facet options.

J=WAT-5141
TEST=manual

demo:https://drive.google.com/file/d/1AfRy4-7Ym1OaO_PQoYIj0tagiPoVyFEN/view?usp=drive_link