fix: duplicate action bar in promo and hero (#822)
We previously were rendering the promo/hero image twice, once on either
side of the content, and hiding one with css. After switching to slots,
this caused a duplicate action bar to appear because Puck saw two Image
components with the same id.

This uses querySelector to find the duplicate action bar and hide it.
~Every once in a while, the bar flashes for a split section, so it's not
perfect but probably is better~


https://github.com/user-attachments/assets/7848f93c-9c7f-4284-be8e-5b0d0e48b203

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>