feat: slot-ify directory (#832)
Some of the spacing in the directory cards changed but I believe it is a
positive change as it removes an odd gap if a location doesn't have a
phone number.

Additionally I changed the vite.config so dts() isn't run during tests,
it was causing my directory tests to occasionally timeout due to hanging
processes.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>