fix: directory meta fields (#839)
The directory meta fields were broken because the default fields are
only populated if `root.props` is undefined and `root.props` was
partially defined in the defaultLayoutData.

This fixed by explicitly adding the title and description fields to the
directory and locator layouts. The field and constantValue are both set
to `""` -- this might change in the future
(https://yext.slack.com/archives/C06A06BCUUF/p1760728433383689?thread_ts=1760724818.768159&cid=C06A06BCUUF)
but we don't have the brand name available right now.